ml/requirements.txt
diffusers>=0.25.0
transformers>=4.36.2
accelerate>=0.25.0
xformers>=0.0.22.post7
torch>=2.1.0
open-clip-torch>=2.24.0
opencv-python>=4.8.0.76
controlnet-aux>=0.0.7
onnxruntime-gpu>=1.16.0
pillow>=10.1.0
einops>=0.7.0
lpips>=0.1.4
scikit-image>=0.22.0
pydantic>=1.10.13
pyyaml>=6.0.1
tqdm>=4.65.0
boto3>=1.28.0
google-api-python-client>=2.108.0
numpy>=1.24.4
scipy>=1.11.4
rich>=13.7.0
